# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RunSafe API — SMB Operations & Compliance SaaS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspace:
    get:
      operationId: getWorkspace
      tags: [workspace]
      summary: Get current user's workspace
      responses:
        "200":
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "404":
          description: No workspace found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createWorkspace
      tags: [workspace]
      summary: Create a new workspace (onboarding)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceBody"
      responses:
        "201":
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
    patch:
      operationId: updateWorkspace
      tags: [workspace]
      summary: Update workspace settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceBody"
      responses:
        "200":
          description: Updated workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"

  # ─── TEAM MEMBERS ────────────────────────────────────────────────
  /team:
    get:
      operationId: getTeamMembers
      tags: [team]
      summary: List team members in the workspace
      responses:
        "200":
          description: List of team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: inviteTeamMember
      tags: [team]
      summary: Invite a team member by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteTeamMemberBody"
      responses:
        "201":
          description: Invitation sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /team/{memberId}:
    patch:
      operationId: updateTeamMember
      tags: [team]
      summary: Update a team member's role
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamMemberBody"
      responses:
        "200":
          description: Updated team member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
    delete:
      operationId: removeTeamMember
      tags: [team]
      summary: Remove a team member
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Member removed

  # ─── SOPs ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── WORKSPACE ───────────────────────────────────────────────
    Workspace:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        industry:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        employeeCount:
          type: ["integer", "null"]
        plan:
          type: string
          enum: [starter, growth, pro]
        ownerClerkId:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - plan
        - ownerClerkId
        - createdAt
    CreateWorkspaceBody:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
        state:
          type: string
        employeeCount:
          type: integer
      required:
        - name
    UpdateWorkspaceBody:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
        state:
          type: string
        employeeCount:
          type: integer
        plan:
          type: string
          enum: [starter, growth, pro]

    # ─── TEAM ────────────────────────────────────────────────────
    TeamMember:
      type: object
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        clerkId:
          type: ["string", "null"]
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [owner, employee]
        status:
          type: string
          enum: [active, invited, inactive]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - workspaceId
        - email
        - name
        - role
        - status
        - createdAt
    InviteTeamMemberBody:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [owner, employee]
      required:
        - email
        - name
    UpdateTeamMemberBody:
      type: object
      properties:
        role:
          type: string
          enum: [owner, employee]
        status:
          type: string
          enum: [active, inactive]

    # ─── SOPs ────────────────────────────────────────────────────
...