# lib/api-zod/src/generated/types/workspace.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RunSafe API — SMB Operations & Compliance SaaS
 * OpenAPI spec version: 0.1.0
 */
import type { WorkspacePlan } from "./workspacePlan";

export interface Workspace {
  id: number;
  name: string;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  employeeCount?: number | null;
  plan: WorkspacePlan;
  ownerClerkId: string;
  createdAt: Date;
}